+++ sketch/emptyp5project.js
@@ -4,10 +4,13 @@ function setup() {
 		window.innerHeight /2
 		);
 
-	canvas.parent('sketch-holder');
+	// canvas.parent('sketch-holder');
 
 }
 
 function draw() {
 	background(200);
+	fill(100); 
+	rect(width/2, height/2, 50, 50);
+	//  
 }