+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- Concatenating js files into concat/concat.js using gulp-concant.
+
+### Changed
+- File structure.
 
 ## [0.1.0] - 2016-12-18
 ### Added